Chapter08/AzureML-Python/requirements.txt
azureml-core
azureml-sdk[explain,automl,notebooks,tensorboard]
configparser==3.8.1
